refactor: moved and renamed project from react-routing
