+++ react-redux/.babelrc
