+++ react-redux/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-routing",
+  "name": "react-redux",
   "version": "1.0.0",
   "description": "",
   "main": "src.index.tsx",
