+++ react-redux/src/base-layout/abstract/_abstract-config.scss
