+++ react-redux/src/base-layout/abstract/_constants.scss
