+++ react-redux/src/base-layout/abstract/_mixins.scss
