+++ react-redux/src/compoments/about/About.module.scss
