+++ react-redux/src/compoments/about/About.tsx
