+++ react-redux/src/compoments/not-found/NotFound.module.scss
