+++ react-redux/src/index.html
@@ -5,7 +5,7 @@
     <meta content='width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0'
           name='viewport' />
     <meta content='ie=edge' http-equiv='X-UA-Compatible' />
-    <title>React Routing</title>
+    <title>React Redux</title>
   </head>
   <body>
     <div id='root'></div>
