+++ react-redux/src/interfaces/IArticle.tsx
