+++ react-redux/src/interfaces/IClasses.tsx
