+++ react-redux/src/interfaces/IData_GET200.tsx
