+++ react-redux/src/interfaces/SortType.tsx
