+++ react-redux/webpack.prod.config.ts
